Price now uses decimals

+++ src/app/Models/Option.php
@@ -19,7 +19,7 @@ class Option extends Model {
 
     protected $casts = [
         'id' => 'integer',
-		'price_modifier' => 'integer',
+		'price_modifier' => 'decimal',
     ];
 
     public $timestamps = false;
