+++ src/database/migrations/2016_01_15_221707_create_products.php
@@ -19,8 +19,8 @@ class CreateProducts extends Migration {
             $table->string('title', 255);
             $table->text('description');
             $table->string('slug', 80);
-            $table->integer('price');
-            $table->integer('price_offer')->nullable();
+            $table->decimal('price', 6, 2);
+            $table->decimal('price_offer', 6, 2)->nullable();
             $table->decimal('weight', 2)->nullable();
             $table->decimal('width', 2)->nullable();
             $table->decimal('height', 2)->nullable();
