+++ src/database/migrations/2016_01_22_221838_create_cart_products.php
@@ -19,7 +19,7 @@ class CreateCartProducts extends Migration {
             $table->foreign('cart_id')->references('id')->on('carts')->onDelete('cascade');
             $table->integer('product_id')->unsigned();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-            $table->integer('price');
+            $table->decimal('price', 6, 2);
         });
     }
